% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piggbyack.R
\name{update_release_timestamp}
\alias{update_release_timestamp}
\title{Write a timestamp for the release}
\source{
\url{https://github.com/nflverse/nflverse-data/R/upload.R#L16}
}
\usage{
update_release_timestamp(temp_dir, tag = tag, timestamp, ...)
}
\description{
Note that this is really only used for the README. The timestamp is also logged as an attribute to the release data.
}
